+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+# 8.2.5 (2024-08-20)
+
+[802c219]: https://github.com/erqk/ng-dynamic-json-form/commit/802c2190f4ea5a09a15fdafd6c2c3e4df11eb3a8
+[18ff62f]: https://github.com/erqk/ng-dynamic-json-form/commit/18ff62fdc4ca4c6fa499605f885bf1c326e58881
+[38ad2ac]: https://github.com/erqk/ng-dynamic-json-form/commit/38ad2ac75c51846e842fe34bc9842dd58c788770
+
+| Commit    | Type | Description                                                |
+| --------- | ---- | ---------------------------------------------------------- |
+| [802c219] | feat | Support custom validator with value.                       |
+| [18ff62f] | fix  | Conditions for validators will only works on the last one. |
+| [38ad2ac] | fix  | Custom action failed to execute.                           |
+
+# 8.2.4 (2024-08-15)
+
+[6d7c9c9]: https://github.com/erqk/ng-dynamic-json-form/commit/6d7c9c9a3b197ae98c7e93e022acdba273045e02
+
+| Commit    | Type | Description                               |
+| --------- | ---- | ----------------------------------------- |
+| [6d7c9c9] | fix  | The comparator is compare the same thing. |
+
+# 8.2.3 (2024-08-10)
+
+[b30179a]: https://github.com/erqk/ng-dynamic-json-form/commit/b30179acac30f2df794793391c03e92b6bf0d866
+
+| Commit    | Type | Description                                           |
+| --------- | ---- | ----------------------------------------------------- |
+| [b30179a] | fix  | Use `CSS.escape()` to escape all invalid charactersw. |
+
 # 8.2.2 (2024-08-06)
 
 [32a14cf]: https://github.com/erqk/ng-dynamic-json-form/commit/32a14cfbf2d1195968b22f201b38f2ac1c53aa68